LargeWidgetBlock added
