+++ Corona Widget/Widget/Widget/MediumWidget.swift
@@ -10,33 +10,18 @@ import SwiftUI
 //MARK: systemMedium Widget
 
 struct mediumWidget : View {
+    // MARK - PROPERTIES
     var data : Countries
+    
+    // MARK - BODY
     var body : some View {
         HStack{
-            PieChart(pieChartData: PieChartData(data: [data.totalActive(), data.TotalDeaths, data.TotalRecovered],colors:[.coronagreen,.coronagrey,.coronayellow]))
+            PieChart(pieChartData: PieChartData(data: [data.totalActive(), data.TotalDeaths, data.TotalRecovered],colors:[.coronayellow,.coronagrey,.coronagreen]))
                 .padding(.all,25)
             Spacer()
-            VStack(alignment:.leading){
-                HStack{
-                    Text(data.emoji())
-                    Text(data.Country)
-                    Spacer()
-                }.font(.system(size: 35))
-                .foregroundColor(.black)
-                .padding(.top,20)
-                VStack(alignment:.leading) {
-                    Text("Total: "+data.TotalConfirmed.cases)
-                        .foregroundColor(.coronapink)
-                    Text("Recovered: "+data.TotalRecovered.cases)
-                        .foregroundColor(.coronagreen)
-                    Text("Deaths: "+data.TotalDeaths.cases)
-                        .foregroundColor(.coronagrey)
-                    Text("Active: "+data.totalActive().cases)
-                        .foregroundColor(.coronayellow)
-                }
-            }.padding(.bottom, 15)
-        }
-    }
+            MediumWidgetBlock(data: data)
+        } //: HSTACK
+    } //: BODY
 }
 
 //END
